contrib.py -> jobutils.py
