+++ dpmModule/jobs/jobutils.py
@@ -1,7 +1,19 @@
+from ..kernel import core
+from ..kernel.graph import DynamicVariableOperation
 from ..item import RootAbyss, Absolab, Arcane
 
-# 스펙 측정시마다 수동으로 변경 필요
+# DamageSkill Duplicator
+def create_auxilary_attack(skill_wrapper, ratio, nametag = '복사'):
+    original_skill = skill_wrapper.skill
+    copial_skill = core.DamageSkill(name = DynamicVariableOperation.reveal_argument(original_skill.name) + nametag,
+        delay = DynamicVariableOperation.wrap_argument(0),
+        damage = original_skill.damage * DynamicVariableOperation.wrap_argument(ratio),
+        hit = original_skill.hit,
+        modifier=original_skill._static_skill_modifier).wrap(core.DamageSkillWrapper)
+    
+    skill_wrapper.onAfter(copial_skill)
 
+# 스펙 측정시마다 수동으로 변경 필요
 def get_weapon_att(WEAPON_NAME, spec = 6000):
     if spec < 5000:
         #파프
